# docs/swagger.yaml
definitions:
...
  controllers.CreateLikeResponse:
    properties:
      message:
        example: like created successfully
        type: string
        x-order: "0"
    type: object
...
  controllers.DeleteCommentResponse:
    properties:
      message:
        example: comment deleted successfully
        type: string
        x-order: "0"
    type: object
  controllers.DeleteLikeResponse:
    properties:
      message:
        example: like deleted successfully
        type: string
        x-order: "0"
    type: object
...
  controllers.EditCommentInput:
    properties:
      comment_text:
        example: sample comment text
        type: string
    required:
    - comment_text
    type: object
  controllers.EditCommentResponse:
    properties:
      comments:
        items:
          $ref: '#/definitions/models.Comment'
        type: array
        x-order: "1"
      message:
        example: comment updated successfully
        type: string
        x-order: "0"
    type: object
...
  controllers.GetCommentLikeResponse:
    properties:
      count:
        example: 1
        type: integer
        x-order: "0"
      rows:
        items:
          $ref: '#/definitions/models.Like'
        type: array
        x-order: "1"
    type: object
...
  models.Comment:
    properties:
      comment_id:
        example: 1
        type: integer
      comment_text:
        example: sample comment text
        type: string
      created_at:
        example: "2030-01-01 00:00:00"
        type: string
      likes:
        items:
          $ref: '#/definitions/models.Like'
        type: array
      review_id:
        example: 1
        type: integer
      user_id:
        example: 1
        type: integer
    type: object
  models.Like:
    properties:
      comment_id:
        example: 1
        type: integer
      created_at:
        example: "2030-01-01 00:00:00"
        type: string
      like_id:
        example: 1
        type: integer
      review_id:
        example: 1
        type: integer
      user_id:
        example: 1
        type: integer
    type: object
...
info:
  contact: {}
paths:
...
  /comments/{comment_id}:
    delete:
      description: Delete a comment
      parameters:
      - description: CommentID
        in: path
        name: comment_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.DeleteCommentResponse'
      security:
      - Bearer: []
      summary: Delete a comment
      tags:
      - Comments
    put:
      description: Update a comment
      parameters:
      - description: CommentID
        in: path
        name: comment_id
        required: true
        type: string
      - description: the body to edit review comment
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.EditCommentInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.EditCommentResponse'
      security:
      - Bearer: []
      summary: Update a comment
      tags:
      - Comments
  /comments/{comment_id}/likes:
    get:
      description: Get all comment likes by comment ID
      parameters:
      - description: CommentID
        in: path
        name: comment_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.GetCommentLikeResponse'
      summary: Get all comment likes by comment ID
      tags:
      - Comments
    post:
      description: Create a comment like by comment ID
      parameters:
      - description: CommentID
        in: path
        name: comment_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/controllers.CreateLikeResponse'
      security:
      - Bearer: []
      summary: Create a comment like by comment ID
      tags:
      - Comments
...
  /likes/{like_id}:
    delete:
      description: Delete a like
      parameters:
      - description: LikeID
        in: path
        name: like_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.DeleteLikeResponse'
      security:
      - Bearer: []
      summary: Delete a like
      tags:
      - Likes
...
securityDefinitions:
  Bearer:
    description: 'Type "Bearer" followed by a space and JWT token (for example: "Bearer
      eyxyz").'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"